refactor(admin): extract shared log-session helpers and loadClusterConfig
Review: HistoricalLogs/_store/session.js mirrored LiveLogs/_store/session.js
(~40% identical) and loadClusterConfig() was byte-identical in both
_router.js files.

The pieces that actually drift on fixes now live once:
- utils/logEvents.js: defaultScopes, addMeta/delMeta, the isFiltered/
  eventsFiltered getter bodies and the UPDATE_FILTERS computation. Both
  stores unify on the sorting addMeta variant (sidebar values stay stably
  ordered while streaming).
- utils/cluster.js (added with the pftest opt-in): both log routers now
  import the cluster-config preload instead of carrying local copies.

Side fix while unifying: LiveLogs LOG_FILTER_ENABLE/DISABLE used plain
assignment on a key that does not exist before the first toggle — not
reactive in Vue 2; both stores now use Vue.set like HistoricalLogs already
did.
